zoos/personalities_zoo/helpers/OCR_converter/requirements.txt
Pillow
pytesseract
